chore: move build helpers into lib/, add requirements.txt, untrack pyc
- New lib/ package houses seo_helpers.py and excel_to_content.py
  (previously at repo root). build.py and tests/ updated to use
  `from lib import ...`. Drops the importlib.util.spec_from_file_location
  dance in build.py since the helper now lives on the import path.

- New requirements.txt lists runtime deps (jinja2, markupsafe, markdown,
  Pillow, openpyxl, beautifulsoup4) with lower-bound pins. The CI
  workflow now installs from this file so the dep list has a single
  source of truth.

- __pycache__/excel_to_content.cpython-39.pyc was tracked in git despite
  __pycache__/ being in .gitignore; un-tracked via `git rm --cached`.

- STRUCTURE.md refreshed: new lib/ tree, stale extract_members.py
  reference removed, dep install step now references requirements.txt.
  CLAUDE.md and CONTRIBUTING.md point at lib/excel_to_content.py.

Verified locally: `python build.py` succeeds end-to-end;
`python -m unittest tests.test_seo_helpers` runs 16/16 tests OK.
