+++ .github/workflows/deploy.yml
@@ -37,7 +37,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install markdown Pillow jinja2 openpyxl beautifulsoup4
+          pip install -r requirements.txt
 
       - name: Run build script
         run: python build.py
