+++ CLAUDE.md
@@ -35,7 +35,7 @@ Data flows from source files → `build.py` → `docs/` (deployed via GitHub Pag
 ```
 contents/member-info.xlsx    ← member source of truth (Excel)
         ↓
-excel_to_content.py          ← auto-generates members JSON + Markdown
+lib/excel_to_content.py      ← auto-generates members JSON + Markdown
         ↓
 contents/structures/*.json   ← page data, members, publications, etc.
 contents/articles/*.md       ← text content (about, member bios)
@@ -43,13 +43,15 @@ templates/*.html             ← Jinja2 templates
 static/                      ← CSS, JS, assets (copied unchanged)
 contents/images/             ← source images (auto-converted to WebP)
         ↓
-    build.py
+    build.py                 ← uses helpers from lib/ (seo_helpers, excel_to_content)
         ↓
     docs/                    ← generated static site (GitHub Pages)
 ```
 
+Python helpers consumed by `build.py` live in `lib/` (`lib/seo_helpers.py`, `lib/excel_to_content.py`); install runtime deps via `pip install -r requirements.txt`.
+
 **`build.py` performs these steps in order:**
-1. **Run `excel_to_content.py`** — reads `contents/member-info.xlsx` and generates `members.json`, per-member JSON files, and member Markdown files (executed at import time)
+1. **Run `lib/excel_to_content.py`** — reads `contents/member-info.xlsx` and generates `members.json`, per-member JSON files, and member Markdown files (executed at import time)
 2. **Load data** — reads all `.json` from `contents/structures/` and `.md` from `contents/articles/` (converted to HTML via `markdown` library); also builds `members_by_id` lookup and a filtered/sorted `home_publications` (only items with `"E3": true` and `status: "published"`)
 3. **Render pages** — iterates `pages.json`, renders each Jinja2 template with the full `structures` dict, writes to `docs/{path}/index.html`
 4. **Render member pages** — for each member in `members.json`, loads `contents/structures/members/{memberId}.json`, auto-populates their publications by matching `pubName` against `authors` in `publications.json`
@@ -78,7 +80,7 @@ All content changes are data-driven — no Python or HTML edits required:
 
 **Adding a new member:**
 1. Add a row to `contents/member-info.xlsx`
-2. Run `python build.py` — `excel_to_content.py` auto-generates:
+2. Run `python build.py` — `lib/excel_to_content.py` auto-generates:
    - `contents/structures/members.json`
    - `contents/structures/members/{webId}.json`
    - `contents/articles/members-{about,position,interest}/{webId}.md`
