+++ STRUCTURE.md
@@ -8,13 +8,17 @@ This repository uses a custom Python static site generator (`build.py`) to build
 
 ```
 NTU-E3-Center.github.io/
-├── build.py                   # Main build script
-├── excel_to_content.py        # Converts member-info.xlsx → JSON + Markdown (called by build.py)
-├── extract_members.py         # Exports all member data to Excel for review/audit
+├── build.py                   # Main build script (CI entry point)
+├── lib/                       # Python helpers consumed by build.py
+│   ├── __init__.py
+│   ├── lib/excel_to_content.py    # Converts member-info.xlsx → JSON + Markdown
+│   └── seo_helpers.py         # Meta-description / markdown-strip / language helpers
+├── requirements.txt           # Runtime dependencies (used by CI and locally)
 ├── contents/                  # Source data, text, and images
 ├── templates/                 # Jinja2 HTML templates
 ├── static/                    # Static assets (copied as-is to output)
-├── docs/                      # Generated output — served by GitHub Pages
+├── tests/                     # pytest suite (import from `lib.…`)
+├── docs/                      # Generated output — served by GitHub Pages (gitignored)
 ├── website/                   # Local Python virtual environment (not committed)
 └── .github/workflows/         # GitHub Actions CI/CD
 ```
@@ -32,7 +36,7 @@ contents/
 ├── member-info.xlsx           # ★ Member source of truth — edit this to add/update members
 ├── structures/
 │   ├── pages.json             # Navigation structure and page paths
-│   ├── members.json           # AUTO-GENERATED by excel_to_content.py — do not edit
+│   ├── members.json           # AUTO-GENERATED by lib/excel_to_content.py — do not edit
 │   ├── publications.json      # Research publications with citation metadata
 │   ├── research.json          # Research topics (researchId-keyed)
 │   ├── about.json             # About section data
@@ -109,7 +113,7 @@ Local Python virtual environment for running `build.py`. Not committed to git.
 ### `.github/workflows/deploy.yml`
 
 GitHub Actions workflow that triggers on every push to the `source` branch:
-1. Installs Python dependencies (`markdown`, `Pillow`, `jinja2`)
+1. Installs Python dependencies via `pip install -r requirements.txt`
 2. Runs `python build.py`
 3. Deploys `docs/` to the `gh-pages` branch via `peaceiris/actions-gh-pages`
 
@@ -119,7 +123,7 @@ GitHub Actions workflow that triggers on every push to the `source` branch:
 
 `build.py` runs five steps in order:
 
-### 1. Run `excel_to_content.py`
+### 1. Run `lib/excel_to_content.py`
 - Reads `contents/member-info.xlsx` (the member source of truth)
 - Generates `contents/structures/members.json`, `contents/structures/members/{webId}.json`, and `contents/articles/members-{about,position,interest}/{webId}.md`
 
@@ -153,7 +157,7 @@ GitHub Actions workflow that triggers on every push to the `source` branch:
 ## Data Flow
 
 ```
-contents/member-info.xlsx   ──── excel_to_content.py ─┐
+contents/member-info.xlsx   ──── lib/excel_to_content.py ─┐
                                                        ↓
 contents/structures/*.json  ─────────────────────────┐ │
 contents/articles/*.md      ─────────────────────────┼─┘
